--- a/samples/BlazorSample/App.razor
+++ b/samples/BlazorSample/App.razor
@@ -1,4 +1,7 @@
-﻿<Router AppAssembly="@typeof(App).Assembly">
+﻿@using LiveChartsCore
+@using ViewModelsSamples
+
+<Router AppAssembly="@typeof(App).Assembly">
     <Found Context="routeData">
         <RouteView RouteData="@routeData" DefaultLayout="@typeof(MainLayout)" />
         <FocusOnNavigate RouteData="@routeData" Selector="h1" />
@@ -10,3 +13,12 @@
         </LayoutView>
     </NotFound>
 </Router>
+
+@code {
+    protected override void OnInitialized()
+    {
+        // this is the recommended place to configure LiveCharts // mark
+        LiveCharts.Configure(config => // mark
+            config.AddLiveChartsAppSettings()); // mark
+    }
+}