--- a/samples/MauiSample/MauiProgram.cs
+++ b/samples/MauiSample/MauiProgram.cs
@@ -1,16 +1,22 @@
 ﻿using LiveChartsCore.SkiaSharpView.Maui;
+using SkiaSharp;
 using SkiaSharp.Views.Maui.Controls.Hosting;
+using LiveChartsCore;
+using ViewModelsSamples;
 
 namespace MauiSample;
 
 public static class MauiProgram
 {
+    public record City(string Name, double Population);
+
     public static MauiApp CreateMauiApp()
     {
         var builder = MauiApp.CreateBuilder();
         _ = builder
             .UseSkiaSharp()
-            .UseLiveCharts()
+            .UseLiveCharts(config => // mark
+                config.AddLiveChartsAppSettings()) // mark
             .UseMauiApp<App>()
             .ConfigureFonts(fonts =>
             {