--- a/samples/UnoPlatformSample/UnoPlatformSample/App.xaml.cs
+++ b/samples/UnoPlatformSample/UnoPlatformSample/App.xaml.cs
@@ -1,6 +1,8 @@
-using Uno.Resizetizer;
+using LiveChartsCore; // mark
+using ViewModelsSamples;
 
 namespace UnoPlatformSample;
+
 public partial class App : Application
 {
     /// <summary>
@@ -17,6 +19,10 @@ public partial class App : Application
 
     protected async override void OnLaunched(LaunchActivatedEventArgs args)
     {
+        // this is the recommended place to configure LiveCharts // mark
+        LiveCharts.Configure(config => // mark
+            config.AddLiveChartsAppSettings()); // mark
+
         var builder = this.CreateBuilder(args)
             // Add navigation support for toolkit controls such as TabBar and NavigationView
             .UseToolkitNavigation()