--- a/samples/ViewModelsSamples/Axes/LabelsFormat2/ViewModel.cs
+++ b/samples/ViewModelsSamples/Axes/LabelsFormat2/ViewModel.cs
@@ -1,24 +1,9 @@
-﻿using SkiaSharp;
-using LiveChartsCore;
-using LiveChartsCore.SkiaSharpView;
-using System;
+﻿using System;
 
 namespace ViewModelsSamples.Axes.LabelsFormat2;
 
 public class ViewModel
 {
-    public ViewModel()
-    {
-        // You must register any non-Latin based font //mark
-        // you can add this code when the app starts to register Chinese characters: // mark
-
-        LiveCharts.Configure(config =>
-            config.HasGlobalSKTypeface(SKFontManager.Default.MatchCharacter('汉')));
-
-        // You can learn more about extra settings at: // mark
-        // https://livecharts.dev/docs/{{ platform }}/{{ version }}/Overview.Installation#configure-themes-fonts-or-mappers-optional // mark
-    }
-
     public double[] Values1 { get; set; } =
         [426, 583, 104];