--- a/samples/WinFormsSample/Program.cs
+++ b/samples/WinFormsSample/Program.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Windows.Forms;
 using LiveChartsCore; // mark
+using ViewModelsSamples;
 
 namespace WinFormsSample;
 
@@ -11,21 +12,11 @@ static class Program
     {
         // this is the recommended place to configure LiveCharts // mark
         LiveCharts.Configure(config => // mark
-        { // mark
-            // configure LiveCharts here... // mark
-        }); // mark;
+            config.AddLiveChartsAppSettings()); // mark
 
         _ = Application.SetHighDpiMode(HighDpiMode.SystemAware);
         Application.EnableVisualStyles();
         Application.SetCompatibleTextRenderingDefault(false);
         Application.Run(new Form1());
     }
-
-    public record City(string Name, double Population);
-}
-
-public class City
-{
-    public string Name { get; set; }
-    public double Population { get; set; }
 }