--- a/src/LiveChartsCore/BarSeries.cs
+++ b/src/LiveChartsCore/BarSeries.cs
@@ -99,7 +99,6 @@ public abstract class BarSeries<TModel, TVisual, TLabel>(
             Fill = v?.Fill ?? Fill,
             Stroke = v?.Stroke ?? Stroke,
             StrokeThickness = (float)MiniatureStrokeThickness,
-            DrawEffect = DrawEffect.Local,
             ClippingBounds = LvcRectangle.Empty,
             Width = (float)MiniatureShapeSize,
             Height = (float)MiniatureShapeSize