--- a/src/LiveChartsCore/CoreAxis.cs
+++ b/src/LiveChartsCore/CoreAxis.cs
@@ -162,14 +162,14 @@ public abstract class CoreAxis<TTextGeometry, TLineGeometry>
     public Paint? NamePaint
     {
         get;
-        set => SetPaintProperty(ref field, value);
+        set => SetPaintProperty(ref field, value, PaintStyle.Text);
     }
 
     /// <inheritdoc cref="IPlane.LabelsPaint"/>
     public Paint? LabelsPaint
     {
         get;
-        set => SetPaintProperty(ref field, value);
+        set => SetPaintProperty(ref field, value, PaintStyle.Text);
     }
 
     /// <inheritdoc cref="IPlane.SeparatorsPaint"/>
@@ -231,7 +231,7 @@ public abstract class CoreAxis<TTextGeometry, TLineGeometry>
     public Paint? CrosshairLabelsPaint
     {
         get;
-        set => SetPaintProperty(ref field, value);
+        set => SetPaintProperty(ref field, value, PaintStyle.Text);
     }
 
     /// <inheritdoc cref="ICartesianAxis.CrosshairLabelsBackground"/>