… docs/overview/1.2.install.md
… samples/ViewModelsSamples/Index.cs
… src/LiveChartsCore/CorePolarAxis.cs
… src/LiveChartsCore/CorePolarLineSeries.cs
… src/LiveChartsCore/CoreScatterSeries.cs
… src/LiveChartsCore/CoreSection.cs
… src/LiveChartsCore/CoreStepLineSeries.cs
… src/LiveChartsCore/Drawing/BaseLabelGeometry.cs
… src/LiveChartsCore/Drawing/BaseVectorGeometry.cs
… src/LiveChartsCore/Drawing/DrawEffect.cs
… src/LiveChartsCore/Drawing/DrawingContext.cs
… src/LiveChartsCore/Drawing/DrawnGeometry.cs
… src/LiveChartsCore/Drawing/Layouts/Layout.cs
… src/LiveChartsCore/Drawing/LvcRectangle.cs
… src/LiveChartsCore/ISeries.cs
… src/LiveChartsCore/Kernel/ChartElement.cs
… src/LiveChartsCore/Kernel/Drawing/Sketch.cs
… src/LiveChartsCore/Kernel/LiveChartsSettings.cs
… src/LiveChartsCore/Kernel/PaintSchedule.cs
… src/LiveChartsCore/Motion/CoreMotionCanvas.cs
… src/LiveChartsCore/Painting/DrawnTask.cs
… src/LiveChartsCore/Painting/MeasureTask.cs
… src/LiveChartsCore/Painting/Paint.cs
… src/LiveChartsCore/Painting/PaintStyle.cs
… src/LiveChartsCore/Series.cs
… src/LiveChartsCore/StrokeAndFillCartesianSeries.cs
… src/LiveChartsCore/VisualElements/BaseAngularTicksVisual.cs
… src/LiveChartsCore/VisualElements/BaseLabelVisual.cs
… src/LiveChartsCore/VisualElements/GeometryVisual.cs
… src/skiasharp/LiveChartsCore.SkiaSharp/Drawing/DrawingExtensions.cs