Addionally to #1915 there is much more the library can do to improve performance; this is especially important on WASM or mobile devices, this PR ensures that:

- SKPaints are cached and reused when possible.
- Shaders are cached when possible.
- Labels shaping is now cached instead of calculated on every frame, this is especially efficient when using not Latin-based fonts, this also improves shaping and labels quality (see #1928).

## Results:

Tested with shaders and Arabic  (requires shaping and RTL alignment), previously to this PR it took ~4.3ms to render a frame, now the same charts takes ~0.79ms to render the same frame, it means that in this case LiveCharts is 5.44 times faster.

About English and more Latin-based fonts, the same previous test but in English takes ~1.44ms to render frame before this PR, and after it takes ~0.8ms which means the library is 1.8 times faster.

**Previously**

![before-performance](https://github.com/user-attachments/assets/0f80b734-c721-454b-9aed-e3f358eceb3f)

**Now**

![after-performance](https://github.com/user-attachments/assets/c12e3f75-a1db-4d85-9cea-304dae0c4623)